+++ config/initializers/container.rb
@@ -162,6 +162,7 @@ module Charkeeper
 
     register('commands.homebrew_context.dnd.add_item') { HomebrewContext::Dnd::Items::AddCommand.new }
     register('commands.homebrew_context.dnd.change_item') { HomebrewContext::Dnd::Items::ChangeCommand.new }
+    register('commands.homebrew_context.dnd.copy_item') { HomebrewContext::Dnd::Items::CopyCommand.new }
 
     register('commands.campaigns_context.add_campaign') { CampaignsContext::AddCampaignCommand.new }
     register('commands.campaigns_context.join_campaign') { CampaignsContext::JoinCampaignCommand.new }